Add user callback to operate on results from QB initial request
At the beginning of a session, QuickBooks sends the complete result
of a CompanyQuery, HostQuery, and PreferencesQuery, as well as
the currently connected company file path, country, and QuickBooks
major/minor version numbers.

This callback can be used to, for example, log/store the company file
and version information.

Shares QBXML parsing and exception handling with `response=`