Add "serialization version" number to config for better backwards compatibility
Request QBXML is generated according to this version of the spec
while response QBXML is parsed according to min_version.

The reason for supporting this is nuanced. Newer versions of QBXML
are backwards compatible with older QBXML versions, meaning that
newer versions of QuickBooks can parse QBXML generated with an older
version number, and a QBXML.rb parser initialized under a newer
version can parse older versions of QBXML, but the inverse is of
course not true. Newer versions of QuickBooks will gleefully
respond to an older-spec QBXML request with elements from a newer
QBXML version, so having the version sent equal to the version
parsed is not a solution for compatibility with anything more than
a single version.

Ideally to cast the widest net, the version number used for
generating will be as low as possible while the version number
accepted for parsing as high as possible.

This commit  makes the supported version effectively a range from
serialization_version to min_version.