Get pending job list from the session during authentication
The result is that should_run? will only be called once per job.
Previously the auth method reconstructed the pending jobs list
from scratch immediately after it was built and stored in the session,
making it possible for conditionally run jobs to produce a different
result, a situation which QBWC is not equipped to handle